+++ src/Products/PloneMeeting/MeetingItem.py
@@ -13,7 +13,6 @@ from collective.contact.plonegroup.config import get_registry_organizations
 from collective.contact.plonegroup.utils import get_all_suffixes
 from collective.contact.plonegroup.utils import get_organization
 from collective.contact.plonegroup.utils import get_organizations
-from collective.contact.plonegroup.utils import get_plone_group
 from collective.contact.plonegroup.utils import get_plone_group_id
 from copy import deepcopy
 from DateTime import DateTime
@@ -4976,7 +4975,12 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         return delay_infos['delay_status'] == 'timed_out' or \
             delay_infos['delay_status_when_stopped'] == 'stopped_timed_out'
 
-    def _updateAdvices(self, invalidate=False, triggered_by_transition=None, inheritedAdviserUids=[]):
+    def _updateAdvices(self,
+                       cfg,
+                       item_state,
+                       invalidate=False,
+                       triggered_by_transition=None,
+                       inheritedAdviserUids=[]):
         '''Every time an item is created or updated, this method updates the
            dictionary self.adviceIndex: a key is added for every advice that needs
            to be given, a key is removed for every advice that does not need to
@@ -5002,9 +5006,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         isDefinedInTool = self.isDefinedInTool()
         if isDefinedInTool:
             self.adviceIndex = PersistentMapping()
-        tool = api.portal.get_tool('portal_plonemeeting')
         plone_utils = api.portal.get_tool('plone_utils')
-        cfg = tool.getMeetingConfig(self)
 
         # check if the given p_triggered_by_transition transition name
         # is the transition that will restart delays
@@ -5212,7 +5214,6 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         # manage PowerAdvisers
         # we will give those groups the ability to give an advice on this item
         # even if the advice was not asked...
-        itemState = self.queryState()
         for org_uid in cfg.getPowerAdvisersGroups():
             # if group already gave advice, we continue
             if org_uid in self.adviceIndex:
@@ -5221,7 +5222,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             # empty because this does not change anything in the UI and adding a
             # user after in the _advisers suffixed Plone group will do things work as expected
             org = get_organization(org_uid)
-            if itemState in org.get_item_advice_states(cfg):
+            if item_state in org.get_item_advice_states(cfg):
                 plone_group_id = get_plone_group_id(org_uid, suffix='advisers')
                 # power advisers get only the right to add the advice, but not to see the item
                 # this must be provided using another functionnality, like power observers or so
@@ -5243,9 +5244,9 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             if 'advice_id' in self.adviceIndex[org_uid]:
                 adviceObj = getattr(self, self.adviceIndex[org_uid]['advice_id'])
             giveReaderAccess = True
-            if itemState not in itemAdviceStates and \
-               itemState not in itemAdviceEditStates and \
-               itemState not in itemAdviceViewStates:
+            if item_state not in itemAdviceStates and \
+               item_state not in itemAdviceEditStates and \
+               item_state not in itemAdviceViewStates:
                 giveReaderAccess = False
                 # in this case, the advice is no more accessible in any way by the adviser
                 # make sure the advice given by groupId is no more editable
@@ -5276,7 +5277,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             # manage delay, add/edit access only if advice is not inherited
             if not self.adviceIsInherited(org_uid):
                 # manage delay-aware advice, we start the delay if not already started
-                if itemState in itemAdviceStates and \
+                if item_state in itemAdviceStates and \
                    self.adviceIndex[org_uid]['delay'] and not \
                    self.adviceIndex[org_uid]['delay_started_on'] and \
                    self.adapted()._adviceDelayMayBeStarted(org_uid):
@@ -5285,7 +5286,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                 # check if user must be able to add an advice, if not already given
                 # check also if the delay is not exceeded, in this case the advice can not be given anymore
                 delayIsNotExceeded = not self._adviceDelayIsTimedOut(org_uid, computeNewDelayInfos=True)
-                if itemState in itemAdviceStates and \
+                if item_state in itemAdviceStates and \
                    not adviceObj and \
                    delayIsNotExceeded and \
                    self.adapted()._adviceIsAddable(org_uid):
@@ -5300,7 +5301,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                     self.adviceIndex[org_uid]['advice_addable'] = True
 
                 # is advice still editable?
-                if itemState in itemAdviceEditStates and \
+                if item_state in itemAdviceEditStates and \
                    delayIsNotExceeded and \
                    adviceObj and \
                    self.adapted()._adviceIsEditable(org_uid):
@@ -5329,18 +5330,18 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                         self.REQUEST.set('mayGiveAdvice', False)
                 # if item needs to be accessible by advisers, it is already
                 # done by self.manage_addLocalRoles here above because it is necessary in any case
-                if itemState in itemAdviceViewStates:
+                if item_state in itemAdviceViewStates:
                     pass
 
                 # make sure there is no 'delay_stopped_on' date if advice still giveable
-                if itemState in itemAdviceStates:
+                if item_state in itemAdviceStates:
                     self.adviceIndex[org_uid]['delay_stopped_on'] = None
                 # the delay is stopped for advices
                 # when the advice can not be given anymore due to a workflow transition
                 # we only do that if not already done (a stopped date is already defined)
                 # and if we are not on the transition that reinitialize delays
                 # and if ever delay was started
-                if itemState not in itemAdviceStates and \
+                if item_state not in itemAdviceStates and \
                    self.adviceIndex[org_uid]['delay'] and \
                    self.adviceIndex[org_uid]['delay_started_on'] and \
                    not isTransitionReinitializingDelays and \
@@ -5669,7 +5670,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             res.append(proposingGroup)
         return res
 
-    def assign_roles_to_group_suffixes(self):
+    def assign_roles_to_group_suffixes(self, cfg, item_state):
         """Method that do the work of assigning relevant roles to
            suffixed groups of an organization depending on current state :
            - suffix '_observers' will have 'Reader' role in every cases;
@@ -5681,30 +5682,25 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
            For unknown states, method _get_corresponding_state_to_assign_local_roles
            will be used to determinate a known configuration to take into ccount"""
         # Add the local roles corresponding to the group managing the item
-        item_state = self.queryState()
         org = self.adapted()._getGroupManagingItem(item_state)
         # in some case like ItemTemplate, we have no proposing group
         if not org:
             return
-
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(self)
         apply_meetingmanagers_access, suffix_roles = compute_item_roles_to_assign_to_suffixes(
             cfg, item_state, org)
 
         # apply local roles to computed suffixes
         for suffix, roles in suffix_roles.items():
-            plone_group = get_plone_group(org.UID(), suffix)
-            if plone_group:
-                self.manage_addLocalRoles(plone_group.id, tuple(roles))
+            # suffix_roles keep only existing suffixes
+            plone_group_id = get_plone_group_id(org.UID(), suffix)
+            self.manage_addLocalRoles(plone_group_id, tuple(roles))
 
         # MeetingManagers get access if item at least validated or decided
         # decided will include states "decided out of meeting"
         # if it is still not decided, it gets full access
         if apply_meetingmanagers_access:
             mmanagers_item_states = ['validated'] + list(cfg.getItemDecidedStates())
-            meeting = self.getMeeting()
-            if item_state in mmanagers_item_states or meeting:
+            if item_state in mmanagers_item_states or self.hasMeeting():
                 mmanagers_group_id = "{0}_{1}".format(cfg.getId(), MEETINGMANAGERS_GROUP_SUFFIX)
                 # 'Reviewer' also on decided item, the WF guard will avoid correct is meeting closed
                 mmanagers_roles = ['Reader', 'Reviewer']
@@ -5729,31 +5725,37 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         # add 'Owner' local role
         self.manage_addLocalRoles(self.owner_info()['id'], ('Owner',))
 
+        tool = api.portal.get_tool('portal_plonemeeting')
+        cfg = tool.getMeetingConfig(self)
+        item_state = self.queryState()
+
         # update suffixes related local roles
-        self.assign_roles_to_group_suffixes()
+        self.assign_roles_to_group_suffixes(cfg, item_state)
 
         # update local roles regarding copyGroups
         isCreated = kwargs.get('isCreated', None)
-        self._updateCopyGroupsLocalRoles(isCreated)
+        self._updateCopyGroupsLocalRoles(isCreated, cfg, item_state)
         # Update advices after updateLocalRoles because updateLocalRoles
         # reinitialize existing local roles
         triggered_by_transition = kwargs.get('triggered_by_transition', None)
         invalidate = kwargs.get('invalidate', False)
         inheritedAdviserUids = kwargs.get('inheritedAdviserUids', [])
-        self._updateAdvices(invalidate=invalidate,
+        self._updateAdvices(cfg,
+                            item_state,
+                            invalidate=invalidate,
                             triggered_by_transition=triggered_by_transition,
                             inheritedAdviserUids=inheritedAdviserUids)
         # Update every 'power observers' local roles given to the
         # corresponding MeetingConfig.powerObsevers
         # it is done on every edit because of 'item_access_on' TAL expression
-        self._updatePowerObserversLocalRoles()
+        self._updatePowerObserversLocalRoles(cfg, item_state)
         # update budget impact editors local roles
         # actually it could be enough to do in in the onItemTransition but as it is
         # always done after updateLocalRoles, we do it here as it is trivial
-        self._updateBudgetImpactEditorsLocalRoles()
+        self._updateBudgetImpactEditorsLocalRoles(cfg, item_state)
         # update group in charge local roles
         # we will give the current groupsInCharge _observers sub group access to this item
-        self._updateGroupsInChargeLocalRoles()
+        self._updateGroupsInChargeLocalRoles(cfg, item_state)
         # manage automatically given permissions
         _addManagedPermissions(self)
         # clean borg.localroles caching
@@ -5771,7 +5773,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                 updateAnnexesAccess(advice)
         # propagate Reader local_roles to sub elements
         # this way for example users that have Reader role on item may view the advices
-        self._propagateReaderAndMeetingManagerLocalRolesToSubObjects()
+        self._propagateReaderAndMeetingManagerLocalRolesToSubObjects(cfg)
         # reindex object security except if avoid_reindex=True and localroles are the same
         avoid_reindex = kwargs.get('avoid_reindex', False)
         if not avoid_reindex or old_local_roles != self.__ac_local_roles__:
@@ -5779,11 +5781,9 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         # return indexes_to_update in case a reindexObject is not done
         return ['getCopyGroups', 'getGroupsInCharge']
 
-    def _propagateReaderAndMeetingManagerLocalRolesToSubObjects(self):
+    def _propagateReaderAndMeetingManagerLocalRolesToSubObjects(self, cfg):
         """Propagate the 'Reader' and 'MeetingManager' local roles to
            sub objects that are blocking local roles inheritance."""
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(self)
         grp_reader_localroles = [
             grp_id for grp_id in self.__ac_local_roles__
             if 'Reader' in self.__ac_local_roles__[grp_id]]
@@ -5794,7 +5794,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                     obj.manage_addLocalRoles(grp_id, ['Reader'])
                 obj.manage_addLocalRoles(meetingmanager_group_id, ['MeetingManager'])
 
-    def _updateCopyGroupsLocalRoles(self, isCreated):
+    def _updateCopyGroupsLocalRoles(self, isCreated, cfg, item_state):
         '''Give the 'Reader' local role to the copy groups
            depending on what is defined in the corresponding meetingConfig.'''
         if not self.isCopiesEnabled():
@@ -5802,11 +5802,8 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         # Check if some copyGroups must be automatically added
         self.addAutoCopyGroups(isCreated=isCreated)
 
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(self)
         # check if copyGroups should have access to this item for current review state
-        itemState = self.queryState()
-        if itemState not in cfg.getItemCopyGroupsStates():
+        if item_state not in cfg.getItemCopyGroupsStates():
             return
         # Add the local roles corresponding to the selected copyGroups.
         # We give the 'Reader' role to the selected groups.
@@ -5815,38 +5812,31 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         for copyGroupId in copyGroupIds:
             self.manage_addLocalRoles(copyGroupId, (READER_USECASES['copy_groups'],))
 
-    def _updatePowerObserversLocalRoles(self):
+    def _updatePowerObserversLocalRoles(self, cfg, item_state):
         '''Give local roles to the groups defined in MeetingConfig.powerObservers.'''
         extra_expr_ctx = _base_extra_expr_ctx(self)
         extra_expr_ctx.update({'item': self, })
-        cfg = extra_expr_ctx['cfg']
         cfg_id = cfg.getId()
-        itemState = self.queryState()
         for po_infos in cfg.getPowerObservers():
-            if itemState in po_infos['item_states'] and \
+            if item_state in po_infos['item_states'] and \
                _evaluateExpression(self,
                                    expression=po_infos['item_access_on'],
                                    extra_expr_ctx=extra_expr_ctx):
                 powerObserversGroupId = "%s_%s" % (cfg_id, po_infos['row_id'])
-                self.manage_addLocalRoles(powerObserversGroupId, (READER_USECASES['powerobservers'],))
+                self.manage_addLocalRoles(powerObserversGroupId,
+                                          (READER_USECASES['powerobservers'],))
 
-    def _updateBudgetImpactEditorsLocalRoles(self):
+    def _updateBudgetImpactEditorsLocalRoles(self, cfg, item_state):
         '''Configure local role for use case 'budget_impact_reviewers' to the corresponding
            MeetingConfig 'budgetimpacteditors' group.'''
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(self)
-        itemState = self.queryState()
-        if itemState not in cfg.getItemBudgetInfosStates():
+        if item_state not in cfg.getItemBudgetInfosStates():
             return
         budgetImpactEditorsGroupId = "%s_%s" % (cfg.getId(), BUDGETIMPACTEDITORS_GROUP_SUFFIX)
         self.manage_addLocalRoles(budgetImpactEditorsGroupId, ('MeetingBudgetImpactEditor',))
 
-    def _updateGroupsInChargeLocalRoles(self):
+    def _updateGroupsInChargeLocalRoles(self, cfg, item_state):
         '''Get the current groupsInCharge and give View access to the _observers Plone group.'''
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(self)
-        itemState = self.queryState()
-        if itemState not in cfg.getItemGroupsInChargeStates():
+        if item_state not in cfg.getItemGroupsInChargeStates():
             return
         groupsInCharge = self.getGroupsInCharge(theObjects=True, includeAuto=True)
         for groupInCharge in groupsInCharge: