+++ src/Products/PloneMeeting/utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from Acquisition import aq_base
 from AccessControl.Permission import Permission
+from Acquisition import aq_base
 from appy.shared.diff import HtmlDiff
 from bs4 import BeautifulSoup
 from collections import OrderedDict
@@ -36,6 +36,7 @@ from plone.autoform.interfaces import WRITE_PERMISSIONS_KEY
 from plone.dexterity.interfaces import IDexterityContent
 from plone.i18n.normalizer.interfaces import IIDNormalizer
 from plone.locking.events import unlockAfterModification
+from plone.memoize import ram
 from Products.Archetypes.event import ObjectEditedEvent
 from Products.CMFCore.permissions import AccessContentsInformation
 from Products.CMFCore.permissions import AddPortalContent
@@ -1706,10 +1707,16 @@ def get_item_validation_wf_suffixes(cfg, org=None, only_enabled=True):
     return suffixes
 
 
+def compute_item_roles_to_assign_to_suffixes_cachekey(method, cfg, item_state, org=None):
+    '''cachekey method for compute_item_roles_to_assign_to_suffixes.'''
+    return cfg.getId(), cfg.modified(), item_state, org and org.UID()
+
+
+@ram.cache(compute_item_roles_to_assign_to_suffixes_cachekey)
 def compute_item_roles_to_assign_to_suffixes(cfg, item_state, org=None):
     """ """
     apply_meetingmanagers_access = True
-    suffix_roles = []
+    suffix_roles = {}
 
     # roles given to item_state are managed automatically
     # it is possible to manage it manually for extra states (coming from wfAdaptations for example)
@@ -1726,7 +1733,10 @@ def compute_item_roles_to_assign_to_suffixes(cfg, item_state, org=None):
         return apply_meetingmanagers_access, suffix_roles
 
     item_val_levels_states = cfg.getItemWFValidationLevels(data='state', only_enabled=True)
+
     # by default, observers may View in every states as well as creators
+    # this way observers have access or it is never the case
+    # and creators have access when state "itemcreated" is disabled
     suffix_roles = {'observers': ['Reader'],
                     'creators': ['Reader']}